[refact] 패키지 구조, css 일부 수정
