+++ src/pages/DetailPage/problem.js
@@ -1,0 +1,5 @@
+function Problem() {
+  return <div></div>;
+}
+
+export default Problem;
