+++ src/pages/DetailPage/problemList.js
@@ -1,0 +1,5 @@
+function ProblemList() {
+  return <div className="border">asd</div>;
+}
+
+export default ProblemList;
